+++ react/components/admin/commissions/CommissionsTable.tsx
@@ -1,5 +1,5 @@
-import type { DataGridColumn } from '@vtex/admin-ui'
 import {
+  useModalState,
   Skeleton,
   useSearchState,
   Search,
@@ -11,10 +11,12 @@ import {
   useDataViewState,
   usePaginationState,
   DataView,
+  useToast,
 } from '@vtex/admin-ui'
+import type { DataGridColumn } from '@vtex/admin-ui'
 import type { FC } from 'react'
 import React, { useState, useEffect } from 'react'
-import { useQuery } from 'react-apollo'
+import { useQuery, useMutation } from 'react-apollo'
 import { useIntl } from 'react-intl'
 import type {
   QueryCommissionsBySkuArgs,
@@ -23,9 +25,11 @@ import type {
 import type { UseSortReturn } from '@vtex/admin-ui/dist/components/DataGrid/hooks/useDataGridSort'
 
 import GET_COMMISSIONS from '../../../graphql/getCommissions.graphql'
+import UPDATE_COMMISSION from '../../../graphql/updateCommission.graphql'
 import { PAGE_SIZE } from '../../../utils/constants'
 import { messages } from '../../../utils/messages'
 import type { CommissionsQueryReturnType } from '../../../typings/tables'
+import EditCommissionModal from './EditCommissionModal'
 
 type TableColumns = {
   id: string
@@ -37,6 +41,9 @@ type TableColumns = {
 const CommissionsTable: FC = () => {
   const intl = useIntl()
   const view = useDataViewState()
+  const modal = useModalState()
+  const showToast = useToast()
+  const [selectedRow, setSelectedRow] = useState<TableColumns>()
   const [sortState, setSortState] = useState<UseSortReturn>()
   const pagination = usePaginationState({
     pageSize: PAGE_SIZE,
@@ -90,6 +97,45 @@ const CommissionsTable: FC = () => {
     },
   })
 
+  const [updateCommissionMutation, { loading: mutationLoading }] = useMutation(
+    UPDATE_COMMISSION,
+    {
+      refetchQueries: [
+        {
+          query: GET_COMMISSIONS,
+          variables: {
+            page: pagination.currentPage,
+            pageSize: PAGE_SIZE,
+            filter: {
+              id: searchState.debouncedValue ?? null,
+            },
+            sorting: sortState?.by
+              ? {
+                  field: sortState.by as CommissionsBySkuSortingField,
+                  order: sortState.order === 'DSC' ? 'DESC' : 'ASC',
+                }
+              : undefined,
+          },
+        },
+      ],
+      awaitRefetchQueries: true,
+      onCompleted: () => {
+        showToast({
+          tone: 'positive',
+          message: intl.formatMessage(messages.editCommissionSuccessMessage),
+        })
+        modal.setVisible(false)
+      },
+      onError: () => {
+        showToast({
+          tone: 'critical',
+          message: intl.formatMessage(messages.editCommissionErrorMessage),
+        })
+        modal.setVisible(false)
+      },
+    }
+  )
+
   const columns: Array<DataGridColumn<TableColumns>> = [
     {
       id: 'skuId',
@@ -116,11 +162,17 @@ const CommissionsTable: FC = () => {
     },
   ]
 
+  const handleRowClick = (item: TableColumns) => {
+    setSelectedRow(item)
+    modal.setVisible(true)
+  }
+
   const dataGridState = useDataGridState<TableColumns>({
     columns,
     length: 3,
     items: data ? data.commissionsBySKU.data : [],
     view,
+    onRowClick: handleRowClick,
   })
 
   // Controls the loading state of the table
@@ -162,6 +214,12 @@ const CommissionsTable: FC = () => {
         />
       </DataViewControls>
       <DataGrid state={dataGridState} />
+      <EditCommissionModal
+        selectedRowId={selectedRow?.id}
+        loading={mutationLoading}
+        updateFunction={updateCommissionMutation}
+        modalState={modal}
+      />
     </DataView>
   )
 }